+++ plugin/src/main/kotlin/net/bnb1/kradle/features/jvm/JavaLintProperties.kt
@@ -1,12 +1,12 @@
 package net.bnb1.kradle.features.jvm
 
-import net.bnb1.kradle.Catalog
 import net.bnb1.kradle.features.Properties
-import net.bnb1.kradle.property
+import net.bnb1.kradle.features.PropertiesDsl
 import org.gradle.api.Project
 
 class JavaLintProperties(project: Project) : Properties(project) {
 
-    val checkstyleConfigFile = property(factory.property("checkstyle.xml"))
-    val checkstyleVersion = property(factory.property(Catalog.Versions.checkstyle))
+    val checkstyle = PropertiesDsl.Builder<CheckstyleProperties>(project)
+        .properties { CheckstyleProperties(it) }
+        .build()
 }