+++ plugin/src/main/kotlin/net/bnb1/kradle/features/jvm/KotlinLintProperties.kt
@@ -1,11 +1,14 @@
 package net.bnb1.kradle.features.jvm
 
-import net.bnb1.kradle.Catalog
 import net.bnb1.kradle.features.Properties
-import net.bnb1.kradle.property
+import net.bnb1.kradle.features.PropertiesDsl
 import org.gradle.api.Project
 
 class KotlinLintProperties(project: Project) : Properties(project) {
 
-    val ktlintVersion = property(factory.property(Catalog.Versions.ktlint))
+    private val ktlintProperties = KtlintProperties(project)
+    val ktlint = PropertiesDsl.Builder<KtlintProperties>(project)
+        .properties(ktlintProperties)
+        .build()
+    val ktlintVersion = ktlintProperties.version
 }